setup react router
