+++ frontend/src/pages/About.tsx
@@ -1,0 +1,5 @@
+const About = () => {
+    return <h1>About</h1>
+}
+
+export default About;
