+++ frontend/src/pages/Help.tsx
@@ -1,0 +1,5 @@
+const Help = () => {
+    return <h1> Help </h1>
+}
+
+export default Help;
