+++ frontend/src/pages/Home.tsx
@@ -1,0 +1,6 @@
+
+const Home = () => {
+    return <h1>Home</h1>
+}
+
+export default Home;
